refactor: split example smoke runner
